refactor: document 패키지로 API Docs 테스트를 위치한다
