+++ src/test/java/com/wootecam/luckyvickyauction/documentation/BuyerAuctionDocument.java
@@ -1,29 +1,83 @@
-package com.wootecam.luckyvickyauction.core.auction.controller;
+package com.wootecam.luckyvickyauction.documentation;
 
+import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyLong;
 import static org.mockito.BDDMockito.given;
+import static org.mockito.BDDMockito.willDoNothing;
 import static org.springframework.restdocs.cookies.CookieDocumentation.cookieWithName;
 import static org.springframework.restdocs.cookies.CookieDocumentation.requestCookies;
 import static org.springframework.restdocs.mockmvc.MockMvcRestDocumentation.document;
 import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.delete;
+import static org.springframework.restdocs.mockmvc.RestDocumentationRequestBuilders.post;
 import static org.springframework.restdocs.payload.PayloadDocumentation.fieldWithPath;
+import static org.springframework.restdocs.payload.PayloadDocumentation.requestFields;
 import static org.springframework.restdocs.payload.PayloadDocumentation.responseFields;
 import static org.springframework.restdocs.request.RequestDocumentation.parameterWithName;
 import static org.springframework.restdocs.request.RequestDocumentation.pathParameters;
+import static org.springframework.restdocs.snippet.Attributes.key;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+import com.wootecam.luckyvickyauction.core.auction.controller.dto.BidRequest;
 import com.wootecam.luckyvickyauction.core.auction.domain.PricePolicy;
+import com.wootecam.luckyvickyauction.core.auction.dto.AuctionSearchCondition;
 import com.wootecam.luckyvickyauction.core.auction.dto.BuyerAuctionInfo;
+import com.wootecam.luckyvickyauction.core.auction.dto.BuyerAuctionSimpleInfo;
 import com.wootecam.luckyvickyauction.core.auction.fixture.BuyerAuctionInfoFixture;
 import com.wootecam.luckyvickyauction.core.member.domain.Role;
 import com.wootecam.luckyvickyauction.core.member.dto.SignInInfo;
-import com.wootecam.luckyvickyauction.documentation.DocumentationTest;
 import jakarta.servlet.http.Cookie;
+import java.time.ZonedDateTime;
+import java.util.ArrayList;
+import java.util.List;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
+import org.springframework.restdocs.payload.JsonFieldType;
 import org.springframework.test.web.servlet.ResultActions;
 
-class BuyerAuctionControllerTest extends DocumentationTest {
+class BuyerAuctionDocument extends DocumentationTest {
+
+    /**
+     * @see <a href="https://github.com/woowa-techcamp-2024/Team7-ELEVEN/issues/89">Github issue</a>
+     */
+    @Nested
+    class 구매자_경매_목록_조회 {
+
+        @Test
+        void 경매_조회_조건을_전달하면_성공적으로_경매_목록을_반환한다() {
+            AuctionSearchCondition condition = new AuctionSearchCondition(0, 2);
+            List<BuyerAuctionSimpleInfo> infos = buyerAuctionSimpleInfosSample();
+            given(auctionService.getBuyerAuctionSimpleInfos(any())).willReturn(infos);
+
+            docsGiven.contentType(MediaType.APPLICATION_JSON_VALUE)
+                    .body(condition)
+                    .when().get("/auctions")
+                    .then().log().all()
+                    .apply(document("auctions/getAuctions/success",
+                            requestFields(
+                                    fieldWithPath("offset").type(JsonFieldType.NUMBER)
+                                            .description("조회를 시작할 순서"),
+                                    fieldWithPath("size").type(JsonFieldType.NUMBER)
+                                            .description("조회할 페이지 크기")
+                                            .attributes(key("constraints").value("최소: 1 ~ 최대: 100"))
+                            )
+                    ))
+                    .statusCode(HttpStatus.OK.value());
+        }
+
+        private List<BuyerAuctionSimpleInfo> buyerAuctionSimpleInfosSample() {
+            List<BuyerAuctionSimpleInfo> infos = new ArrayList<>();
+
+            for (long i = 1; i <= 2; i++) {
+                BuyerAuctionSimpleInfo simpleInfo = new BuyerAuctionSimpleInfo(i, "쓸만한 경매품 " + i, i * 2000,
+                        ZonedDateTime.now(), ZonedDateTime.now().plusMinutes(30L));
+                infos.add(simpleInfo);
+            }
+
+            return infos;
+        }
+    }
 
     @Nested
     class 구매자_경매_취소_API {
@@ -97,12 +151,10 @@ class BuyerAuctionControllerTest extends DocumentationTest {
                             )
                     ))
                     .statusCode(200);
-
         }
 
         @Test
         void PercentagePolicy_경매_조회시() {
-            // given
             Long auctionId = 1L;
             BuyerAuctionInfo response = BuyerAuctionInfoFixture.create()
                     .auctionId(auctionId)
@@ -111,9 +163,7 @@ class BuyerAuctionControllerTest extends DocumentationTest {
             given(auctionService.getBuyerAuction(auctionId))
                     .willReturn(response);
 
-            // when
-            docsGiven
-                    .accept(MediaType.APPLICATION_JSON_VALUE)
+            docsGiven.accept(MediaType.APPLICATION_JSON_VALUE)
                     .contentType(MediaType.APPLICATION_JSON_VALUE)
                     .when()
                     .get("/auctions/{auctionId}", auctionId.toString())
@@ -136,8 +186,41 @@ class BuyerAuctionControllerTest extends DocumentationTest {
                             )
                     ))
                     .statusCode(200);
-
         }
     }
 
+    @Nested
+    class 구매자_경매_입찰 {
+
+        @Test
+        void 경매_입찰을_성공하면_OK응답을_반환한다() throws Exception {
+            String auctionId = "1";
+            BidRequest bidRequest = new BidRequest(10000L, 20L);
+            SignInInfo buyerInfo = new SignInInfo(1L, Role.BUYER);
+            willDoNothing().given(paymentService)
+                    .process(any(SignInInfo.class), anyLong(), anyLong(), anyLong(), any(ZonedDateTime.class));
+            given(authenticationContext.getPrincipal()).willReturn(buyerInfo);
+
+            mockMvc.perform(post("/auctions/{auctionId}/bids", auctionId)
+                            .contentType(MediaType.APPLICATION_JSON_VALUE)
+                            .sessionAttr("signInMember", buyerInfo)
+                            .cookie(new Cookie("JSESSIONID", "sessionId"))
+                            .content(objectMapper.writeValueAsString(bidRequest)))
+                    .andDo(document("auctions/bids/success",
+                            requestCookies(
+                                    cookieWithName("JSESSIONID").description("세션 ID")
+                            ),
+                            pathParameters(
+                                    parameterWithName("auctionId").description("입찰한 경매의 ID")
+                            ),
+                            requestFields(
+                                    fieldWithPath("price").type(JsonFieldType.NUMBER)
+                                            .description("입찰을 희망하는 가격"),
+                                    fieldWithPath("quantity").type(JsonFieldType.NUMBER)
+                                            .description("입찰을 희망하는 수량")
+                            )
+                    ))
+                    .andExpect(status().isOk());
+        }
+    }
 }
